requirements-dev.txt
# Development dependencies

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0
